requirements.txt
pillow
opencv-python
torch
torchvision
